+++ test/misc.t
@@ -36,3 +36,15 @@ Unicode surrogate should not crash the parser
         (location ((start ((line 1) (column 0))) (stop ((line 1) (column 8)))))
         (value (String "\\ud83d")) (kind ()))))))
    (type_ignores ()))
+
+Unicode escape sequence "\N" is not supported
+  $ echo '"\N{HORIZONTAL ELLIPSIS}"' | parse module -
+  ((body
+    ((Expr
+      (location ((start ((line 1) (column 0))) (stop ((line 1) (column 25)))))
+      (value
+       (Constant
+        (location
+         ((start ((line 1) (column 0))) (stop ((line 1) (column 25)))))
+        (value (String "\\N{HORIZONTAL ELLIPSIS}")) (kind ()))))))
+   (type_ignores ()))